Create an actual service for the processor
