+++ gardening/common/src/commonMain/kotlin/com/datasite/poc/garden/dto/User.kt
@@ -1,0 +1,9 @@
+package com.datasite.poc.garden.dto
+
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class User(
+    val id: String,
+    val name: String,
+)
