+++ gardening/service/src/main/resources/application.yml
