+++ processor/service/build.gradle.kts
@@ -1,0 +1,68 @@
+import org.jetbrains.kotlin.gradle.tasks.KotlinCompile
+
+plugins {
+	id("org.springframework.boot")
+	id("io.spring.dependency-management")
+	kotlin("jvm")
+	kotlin("plugin.spring")
+}
+
+java.sourceCompatibility = JavaVersion.VERSION_11
+
+repositories {
+	mavenCentral()
+}
+
+extra["springCloudVersion"] = "Hoxton.SR8"
+extra["testcontainersVersion"] = "1.14.3"
+
+dependencies {
+	implementation(project(":processor:processor-common"))
+	implementation(project(":gardening:gardening-common"))
+	implementation(project(":auditing:auditing-common"))
+
+	implementation("org.springframework.cloud:spring-cloud-stream")
+	implementation("org.springframework.cloud:spring-cloud-stream-binder-kafka-streams")
+
+	implementation("com.fasterxml.jackson.module:jackson-module-kotlin")
+
+	implementation("org.jetbrains.kotlin:kotlin-reflect")
+	implementation("org.jetbrains.kotlin:kotlin-stdlib-jdk8")
+	implementation("org.jetbrains.kotlinx:kotlinx-coroutines-reactor:1.4.0")
+	implementation("org.jetbrains.kotlinx:kotlinx-coroutines-reactive:1.4.0")
+	implementation("org.jetbrains.kotlinx:kotlinx-coroutines-core:1.4.0")
+
+	// Unit Tests
+	testImplementation("org.springframework.boot:spring-boot-starter-test") {
+		exclude(group = "org.junit.vintage", module = "junit-vintage-engine")
+	}
+	testImplementation("io.projectreactor:reactor-test")
+	testImplementation("org.springframework.kafka:spring-kafka-test")
+	testImplementation("org.testcontainers:junit-jupiter")
+	testImplementation("org.testcontainers:kafka")
+	testImplementation("org.testcontainers:postgresql")
+}
+
+dependencyManagement {
+	imports {
+		mavenBom("org.springframework.cloud:spring-cloud-dependencies:${property("springCloudVersion")}")
+		mavenBom("org.testcontainers:testcontainers-bom:${property("testcontainersVersion")}")
+	}
+}
+
+tasks.withType<Test> {
+	useJUnitPlatform()
+}
+
+tasks.withType<KotlinCompile> {
+	kotlinOptions {
+		freeCompilerArgs = listOf("-Xjsr305=strict")
+		jvmTarget = "11"
+	}
+}
+
+tasks.processResources.configure {
+	from(tasks.getByPath(":reporting:reporting-ui:jsBundle")) {
+		into("static")
+	}
+}
