+++ processor/service/src/main/kotlin/com/datasite/poc/garden/event/DatasiteGardenEventApplication.kt
@@ -1,0 +1,11 @@
+package com.datasite.poc.garden.event
+
+import org.springframework.boot.autoconfigure.SpringBootApplication
+import org.springframework.boot.runApplication
+
+@SpringBootApplication
+class DatasiteGardenEventApplication
+
+fun main(args: Array<String>) {
+    runApplication<DatasiteGardenEventApplication>(*args)
+}
