+++ processor/service/src/main/resources/application.yml
@@ -1,0 +1,20 @@
+server:
+  port: 0 # Doesn't matter what port
+
+spring:
+  kafka:
+    bootstrapServers: localhost:9092
+
+spring.cloud.stream.bindings:
+  mongoGardenOpLog:
+    destination: mongo.datasite-poc.gardens
+  mongoGardenTable:
+    destination: mongo.garden.table
+  gardenAuditLog:
+    destination: audit.garden.event
+  gardenEnhancedAuditLog:
+    destination: audit.garden.event.enhanced
+
+logging:
+  level:
+    org.apache.kafka: WARN
