--- service/build.gradle.kts
@@ -1,64 +1,0 @@
-import org.jetbrains.kotlin.gradle.tasks.KotlinCompile
-
-plugins {
-	id("org.springframework.boot")
-	id("io.spring.dependency-management")
-	kotlin("jvm")
-	kotlin("plugin.spring")
-}
-
-java.sourceCompatibility = JavaVersion.VERSION_11
-
-repositories {
-	mavenCentral()
-}
-
-extra["testcontainersVersion"] = "1.14.3"
-
-dependencies {
-	implementation(project(":auditing:auditing-common"))
-
-	implementation("org.springframework.boot:spring-boot-starter-webflux")
-	implementation("org.springframework.boot:spring-boot-starter-actuator")
-
-	implementation("com.fasterxml.jackson.module:jackson-module-kotlin")
-
-	implementation("org.springframework.boot:spring-boot-starter-data-mongodb-reactive")
-
-	implementation("org.apache.kafka:kafka-streams")
-	implementation("org.apache.kafka:kafka-clients")
-	implementation("org.springframework.kafka:spring-kafka")
-
-	implementation("org.jetbrains.kotlin:kotlin-reflect")
-	implementation("org.jetbrains.kotlin:kotlin-stdlib-jdk8")
-	implementation("org.jetbrains.kotlinx:kotlinx-coroutines-reactor:1.4.0")
-	implementation("org.jetbrains.kotlinx:kotlinx-coroutines-reactive:1.4.0")
-	implementation("org.jetbrains.kotlinx:kotlinx-coroutines-core:1.4.0")
-
-	// Unit Tests
-	testImplementation("org.springframework.boot:spring-boot-starter-test") {
-		exclude(group = "org.junit.vintage", module = "junit-vintage-engine")
-	}
-	testImplementation("io.projectreactor:reactor-test")
-	testImplementation("org.springframework.kafka:spring-kafka-test")
-	testImplementation("org.testcontainers:junit-jupiter")
-	testImplementation("org.testcontainers:kafka")
-	testImplementation("org.testcontainers:mongodb")
-}
-
-dependencyManagement {
-	imports {
-		mavenBom("org.testcontainers:testcontainers-bom:${property("testcontainersVersion")}")
-	}
-}
-
-tasks.withType<Test> {
-	useJUnitPlatform()
-}
-
-tasks.withType<KotlinCompile> {
-	kotlinOptions {
-		freeCompilerArgs = listOf("-Xjsr305=strict")
-		jvmTarget = "11"
-	}
-}
